+++ vscode-extension/src/configmanager/ALBuildConfigurationNaming.ts
@@ -1,0 +1,6 @@
+export enum ALBuildConfigurationNaming {
+    none = 0,
+    appNameJson = 1,
+    nameAppJson = 2,
+    appJsonName = 3
+}