+++ vscode-extension/src/const/settingsConst.ts
@@ -1,0 +1,6 @@
+export class SettingsConst {
+    static readonly ExtensionSettingsSection: string = 'alOutline';
+    
+    static readonly ActiveBuildConfiguration: string = 'activeBuildConfiguration';
+    static readonly BuildConfigurationNaming: string = 'buildConfigurationNaming';
+}