+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 13.69.0
+ - Multiple build configurations support (multiple app.json files)
+   - Status bar button to switch between configurations
+   - Naming rules for app.json files defined in `alOutline.buildConfigurationNaming` setting
+
 ## 13.68.0
  - ToolTips functionality changes
    - new `alOutline.fieldToolTipsLocation` setting, avaliable values are